Change gene bulk upload test order and expected results counts
This is done due to the order in which test classes are run.
The test for bulk upload of MOD example genes has been moved to
be the first test run in that class as the same load is required
in the disease annotation bulk test class (in order for genes to
be present and annotations not skipped).

A better approach might be to implement manual ordering of test
class execution (I couldn't get this to work with Quarkus) or to
start with an empty database for each class (would increase test
run time).